refactor: move ValueProposition into own component folder
